Feat get and post methods

+++ concessionaria/src/main/java/com/concessionaria/anunciante/AnuncianteRepository.java
@@ -2,5 +2,5 @@ package com.concessionaria.anunciante;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnuncianteRepository extends JpaRepository<Anunciante,Long>
-{}
+public interface AnuncianteRepository extends JpaRepository<Anunciante, Integer> {
+}
